+++ integration_tests/assets/confd_data/mock-wazo-confd.py
@@ -65,7 +65,12 @@ def extensions_features():
 
 @app.route('/{}/users/<user_uuid>'.format(API_VERSION), methods=['GET'])
 def user_uuid_get(user_uuid):
-    return jsonify({'id': 123})
+    return jsonify({'id': user_uuid, 'uuid': user_uuid, 'lines': [{'name': 'line-123'}]})
+
+
+@app.route('/{}/users/<user_uuid>/services/<service_name>'.format(API_VERSION), methods=['PUT'])
+def user_service_put(user_uuid, service_name):
+    return '', 204
 
 
 if __name__ == "__main__":