+++ wazo_phoned/plugin_helpers/client/schema.py
@@ -1,4 +1,4 @@
-# Copyright 2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2019-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo.mallow import fields
@@ -13,3 +13,8 @@ class LookupSchema(UserUUIDSchema):
     term = fields.String(required=True)
     limit = fields.Integer(missing=None)
     offset = fields.Integer(missing=0)
+
+
+class UserServiceSchema(Schema):
+    user_uuid = fields.String(required=True)
+    enabled = fields.Boolean(missing=None)