+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 Changelog
 =========
 
+20.06
+-----
+
+* Added the DND status synchronization for Yealink phones
+* Added the following endpoint:
+
+  * GET `/0.1/yealink/user_service/dnd`
+
 20.03
 -----
 
 * Added a plugin to handle BLF events.
-
 
 19.17
 -----